add CudaEvent

+++ cuda_event.h
@@ -8,7 +8,7 @@ namespace mtk{
 		std::map<std::string,cudaEvent_t> events_map;
 	public:
 		CudaEvent* createEvent(std::string event_name);
-		unsigned int elapsedTime(std::string start_event,std::string stop_event);
+		float elapsedTime(std::string start_event,std::string stop_event);
 		void recordEvent(std::string event_name);
 	};
 }
